+++ temporalio/contrib/gcp/cloud_run/__init__.py
@@ -1,0 +1,1 @@
+"""Google Cloud Run integrations for Temporal."""